move sql statement schema holder class
